# agent-tuteur-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Tuteur Sénégal — API",
    "description": "Agent tuteur pédagogique RAG pour le programme scolaire sénégalais.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/quiz": {
      "post": {
        "tags": [
          "quiz"
        ],
        "summary": "Generate Quiz",
        "description": "Sert une question d'évaluation **tirée du corpus** sur une compétence.\n\nL'absence de question n'est **pas** une erreur HTTP : c'est le résultat\nhonnête d'un chapitre dont la section d'auto-évaluation est vide, absente,\nou dont les QCM ne portent pas de clé de correction. L'interface affiche\nalors ``instructions`` — jamais un questionnaire de remplissage, jamais une\nquestion inventée pour l'occasion.",
        "operationId": "generate_quiz_api_quiz_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/quiz/answer": {
      "post": {
        "tags": [
          "quiz"
        ],
        "summary": "Answer Quiz",
        "description": "Corrige une réponse, met à jour la maîtrise, attribue les badges mérités.",
        "operationId": "answer_quiz_api_quiz_answer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizCorrectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BadgeOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "earned_at": {
            "type": "string",
            "title": "Earned At"
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "earned_at"
        ],
        "title": "BadgeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MasteryEntryOut": {
        "properties": {
          "competence": {
            "type": "string",
            "title": "Competence"
          },
          "chapitre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapitre"
          },
          "mastery_score": {
            "type": "number",
            "title": "Mastery Score"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "successes": {
            "type": "integer",
            "title": "Successes"
          },
          "statut": {
            "type": "string",
            "title": "Statut"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          }
        },
        "type": "object",
        "required": [
          "competence",
          "mastery_score",
          "attempts",
          "successes",
          "statut"
        ],
        "title": "MasteryEntryOut"
      },
...
      "QuizAnswerRequest": {
        "properties": {
          "quiz_token": {
            "type": "string",
            "minLength": 1,
            "title": "Quiz Token"
          },
          "answer": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Answer"
          },
          "student_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          }
        },
        "type": "object",
        "required": [
          "quiz_token",
          "answer"
        ],
        "title": "QuizAnswerRequest"
      },
      "QuizChoiceOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "title": "QuizChoiceOut"
      },
      "QuizCorrectionOut": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "given_answer": {
            "type": "string",
            "title": "Given Answer"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "mastery": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MasteryEntryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "badges": {
            "items": {
              "$ref": "#/components/schemas/BadgeOut"
            },
            "type": "array",
            "title": "Badges"
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "correct_answer",
          "given_answer",
          "explanation",
          "score"
        ],
        "title": "QuizCorrectionOut"
      },
      "QuizOut": {
        "properties": {
          "competence": {
            "type": "string",
            "title": "Competence"
          },
          "quiz_type": {
            "type": "string",
            "title": "Quiz Type"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "question": {
            "type": "string",
            "title": "Question",
            "default": ""
          },
          "choices": {
            "items": {
              "$ref": "#/components/schemas/QuizChoiceOut"
            },
            "type": "array",
            "title": "Choices"
          },
          "quiz_token": {
            "type": "string",
            "title": "Quiz Token",
            "default": ""
          },
          "instructions": {
            "type": "string",
            "title": "Instructions",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "competence",
          "quiz_type",
          "available"
        ],
        "title": "QuizOut",
        "description": "Quiz remis à l'élève. **Sans la bonne réponse** — voir ``quiz_token``."
      },
      "QuizRequest": {
        "properties": {
          "competence": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Competence"
          },
          "quiz_type": {
            "type": "string",
            "enum": [
              "qcm",
              "vrai_faux"
            ],
            "title": "Quiz Type",
            "default": "qcm"
          },
          "curriculum_context": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Curriculum Context"
          }
        },
        "type": "object",
        "title": "QuizRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}